Feat/multi group snapshot (#42)
* (fix) multi-group snapshot with memoryStore

* (fix) multi-group snapshot with rocksDBStore

* (fix) add multi sst snapshot benchmark

* (fix) add meta check on load snapshot

* (fix) add meta check on load snapshot

* (fix) add timer metric with memory snapshot

* (fix) typo

* (fix) log

* (bugfix) fencing token isolation

* (bugfix) rename local variable

* (bugfix) fencing token isolation by region

* (fix) typo

* (fix) region info visibility

* (feat) add comment

* (fix) delete existing data on rocksdb restart, because raft will play them back

* (fix) delete existing data on rocksdb restart, because raft will play them back

* (fix) add restart test #86

* (fix) add restart test #86

* (fix) update restart test

* (fix) add error log on splitting fail

* (fix) requires

* (fix) remove useless code

* (fix) code refactoring with snapshot

* (fix) fix bad name

* (fix) typo

+++ jraft-rheakv/rheakv-core/src/test/java/com/alipay/sofa/jraft/rhea/storage/rhea/RheaKVTestCluster.java
@@ -52,11 +52,15 @@ public class RheaKVTestCluster {
     private List<RheaKVStore>     stores         = new CopyOnWriteArrayList<>();
 
     protected void start(final StorageType storageType) throws Exception {
-        deleteFiles();
+        start(storageType, true);
+    }
+
+    protected void start(final StorageType storageType, final boolean deleteFiles) throws Exception {
+        if (deleteFiles) {
+            deleteFiles();
+        }
         for (final String c : CONF) {
-            final ObjectMapper mapper = new ObjectMapper(new YAMLFactory());
-            final InputStream in = RheaKVTestCluster.class.getResourceAsStream(c);
-            final RheaKVStoreOptions opts = mapper.readValue(in, RheaKVStoreOptions.class);
+            final RheaKVStoreOptions opts = readOpts(c);
             opts.getStoreEngineOptions().setStorageType(storageType);
             final RheaKVStore rheaKVStore = new DefaultRheaKVStore();
             if (rheaKVStore.init(opts)) {
@@ -71,11 +75,25 @@ public class RheaKVTestCluster {
     }
 
     protected void shutdown() throws Exception {
+        shutdown(true);
+    }
+
+    protected void shutdown(final boolean deleteFiles) throws Exception {
         for (final RheaKVStore store : stores) {
             store.shutdown();
         }
-        deleteFiles();
+        stores.clear();
+        if (deleteFiles) {
+            deleteFiles();
+        }
         LOG.info("RheaKVTestCluster shutdown complete");
+    }
+
+    private RheaKVStoreOptions readOpts(final String conf) throws IOException {
+        final ObjectMapper mapper = new ObjectMapper(new YAMLFactory());
+        try (final InputStream in = RheaKVTestCluster.class.getResourceAsStream(conf)) {
+            return mapper.readValue(in, RheaKVStoreOptions.class);
+        }
     }
 
     protected RheaKVStore getRandomLeaderStore() {
